Turing Machine Implementation
Implementation of the Turing Machine. Program reads in Turing Machine specifications from a text file, and reads input from the user. If a Turing Machine provided finished on accept state the input is accepted. If it halts on the reject state, it is rejected. It is possible that Turing Machine can enter an infinite loop.